// data/dto/esaf-account-view.dto.ts
export type EsafLightWeightAccountViewRequestDto = {
  args0: {
    bankCode: number;
    channel: string;
    externalReferenceNo: string;
    serviceCode: string;
    transactionBranch: number;
    userId: string;
    postingDateText: string;
  };
  args1: {
    codAcctNo: string;
  };
};

export type EsafLightWeightAccountViewResponseDto = {
  postingDate?: string;
  transactionStatus?: {
    FCYHangeHandlingApplied?: boolean;
    errorCode?: string;
    extendedReply?: Record<string, unknown>;
    externalReferenceNo?: string;
    isOverriden?: boolean;
    isServiceChargeApplied?: boolean;
    replyCode?: number;
    spReturnValue?: number;
  };
  xfaceAccountDetailsforCustomerDTO?: Array<{
    DPD?: number;
    NPA?: string;
    accountCRR?: string;
    accountCloseDate?: string;
    accountId?: string;
    accountTitle?: string;
    acctAddr1?: string;
    acctAddr2?: string;
    acctAddr3?: string;
    acctCity?: string;
    acctCountry?: string;
    acctEmailID?: string;
    acctMobileNumber?: string;
    acctState?: string;
    acctZipCode?: string;
    amtShortfallMab?: number;
    availableBalanace?: number;
    balPrincipal?: number;
    balanceBook?: number;
    branchCode?: string;
    branchName?: string;
    classification?: string;
    codDepNo?: number;
    currencyCode?: number;
    currencyShortName?: string;
    currentStatus?: string;
    customerRelationship?: string;
    datAcctOpen?: string;
    datLoa?: string;
    datMaturity?: string;
    datPoa?: string;
    datValLoa?: string;
    datValPoa?: string;
    flgBlocked?: string;
    flgChqMandate?: string;
    flgChqbk?: string;
    flgEmployeeAccount?: string;
    flgPassbook?: string;
    flgRestricted?: string;
    flgSweepin?: string;
    flgSweepout?: string;
    flgUpi?: string;
    holdBalance?: number;
    iban?: string;
    ifscCode?: string;
    micrCode?: string;
    minimumBalance?: number;
    modeOfOperation?: string;
    moduleCode?: string;
    originalBalance?: number;
    outstandingAmount?: number;
    overDraftLimit?: number;
    productCategory?: string;
    productCode?: string;
    productName?: string;
    ratInt?: number;
    sanctionAmount?: number;
    scheduleDueAmount?: number;
    tenure?: string;
    unclearFunds?: number;
    universalCRR?: string;
    ytdIntReceived?: number;
  }>;
  xfaceCustomerResponseDTO?: {
    aadhaarReferenceNo?: string;
    bankShortName?: string;
    city?: string;
    country?: string;
    customerFullName?: string;
    customerId?: number;
    emailAddress?: string;
    flgBlocked?: string;
    line1?: string;
    line2?: string;
    line3?: string;
    mobileNumber?: string;
    pan?: string;
    shortName?: string;
    state?: string;
    zip?: string;
  };
};